Some PR review comments, and system cleanup
